Extract resolvePath into a unit-tested thingamabob
